[ refactor ] factor out code for telemetry process
